Move Copy file from Bundle to FileOperationUtils Use Copy instead of Directory.Move since it sometimes throws
